backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Security Scanning
semgrep==1.55.0

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.41

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0

# AI Integration
anthropic==0.7.8
openai==1.3.8
